edit service

+++ driver/mysql/user.go
@@ -2,6 +2,7 @@ package mysql
 
 import (
 	"errors"
+	"ticketing/helper/encrypt"
 	"ticketing/model/domain"
 
 	"gorm.io/gorm"
@@ -17,7 +18,9 @@ func NewDBUserService(db *gorm.DB) *DBUserService {
 	}
 }
 
-func (us *DBUserService) Save(user domain.User) (domain.User, error) {
+func (us *DBUserService) Add(user domain.User) (domain.User, error) {
+	pw, _ := encrypt.Hash(user.Password)
+	user.Password = pw
 	tx := us.db.Save(&user)
 	err := tx.Error
 	return user, err
@@ -40,10 +43,41 @@ func (us *DBUserService) Get(id uint) (domain.User, error) {
 	return user, err
 }
 
-func (us *DBUserService) Delete(id uint) (domain.User, error) {
+func (us *DBUserService) Update(id uint, user domain.User, jwtID uint) (domain.User, error) {
+	userDB, err := us.Get(id)
+	if err != nil {
+		return user, err
+	}
+	if userDB.ID != jwtID {
+		return domain.User{}, errors.New("forbidden")
+	}
+
+	if user.Name != "" {
+		userDB.Name = user.Name
+	}
+	if user.Email != "" {
+		userDB.Email = user.Email
+	}
+	if user.Password != "" {
+		pw, _ := encrypt.Hash(user.Password)
+		userDB.Password = pw
+	}
+	if user.Phone != "" {
+		userDB.Phone = user.Phone
+	}
+
+	tx := us.db.Save(&userDB)
+	err = tx.Error
+	return user, err
+}
+
+func (us *DBUserService) Delete(id uint, jwtID uint) (domain.User, error) {
 	user, err := us.Get(id)
 	if err != nil {
 		return user, err
+	}
+	if id != jwtID {
+		return domain.User{}, errors.New("forbidden")
 	}
 	tx := us.db.Delete(&user)
 	err = tx.Error
